// src/app/shared/enums/activity-log.action.enum.ts
export enum CallanActivityLogActionEnum {
    BALANCE_SPEND = 'BALANCE_SPEND',
    BALANCE_REFUND = 'BALANCE_REFUND',
    BALANCE_CHANGE = 'BALANCE_CHANGE',
    LESSON_EVENT_CREATE = 'LESSON_EVENT_CREATE',
    LESSON_EVENT_STATUS_CHANGE = 'LESSON_EVENT_STATUS_CHANGE',
    LESSON_EVENT_AUTO_STATUS_CHANGE = 'LESSON_EVENT_AUTO_STATUS_CHANGE',
    LESSON_EVENT_TEACHER_ASSIGNMENT = 'LESSON_EVENT_TEACHER_ASSIGNMENT',
    CUSTOMER_CREATE = 'CUSTOMER_CREATE',
    CUSTOMER_UPDATE = 'CUSTOMER_UPDATE',
}